LayoutTests/accessibility/ax-thread-text-apis/text-marker-sentence-nav.html

<!DOCTYPE HTML><!-- webkit-test-runner [ runSingly=true AccessibilityThreadTextApisEnabled=true ] -->
<!-- This is a modern-style rewrite of mac/text-marker-sentence-nav.html, and should replace it after AccessibilityThreadTextApisEnabled is enabled by default. -->
<html>
<head>
<meta charset="utf-8"> 
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
</head>
<body id="body">

<div id="text1" tabindex="0">
This i<span>s a sen</span>tence,
right? Yes!
</div>

<div id="text2">
test audio <audio controls><source src="test.mp3" type="audio/mpeg"></audio>file.
</div>

<div id="text3">
巧克力
是食物吗? 是的。
</div>

<div id="text3a">
كيف حالك؟
</div>

<pre id="text4">
both   spaces
line breaks. Right?
</pre>

<div id="text5">
This is my first sentence.<br><br>
This is my second sentence. This is my third sentence.
This is my fourth sentence.
</div>

<script>
var output = "This tests that sentence navigation is working correctly.\n\n"

if (window.accessibilityController) {

    // Check that we can get the paragraph range with span tag.
    var text = accessibilityController.accessibleElementById("text1");
    var textMarkerRange = text.textMarkerRangeForElement(text);
    var startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
    var currentMarker = advanceAndVerify(startMarker, 1, text);
    
    // Audio tag.
    text = accessibilityController.accessibleElementById("text2");
    textMarkerRange = text.textMarkerRangeForElement(text);
    startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
    currentMarker = advanceAndVerify(startMarker, 1, text);
    
    // Multi-languages.
    text = accessibilityController.accessibleElementById("text3");
    textMarkerRange = text.textMarkerRangeForElement(text);
    startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
    currentMarker = advanceAndVerify(startMarker, 1, text);
    
    text = accessibilityController.accessibleElementById("text3a");
    textMarkerRange = text.textMarkerRangeForElement(text);
    startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
    currentMarker = advanceAndVerify(startMarker, 1, text);
    
    // Check the case with pre tag.
    text = accessibilityController.accessibleElementById("text4");
    textMarkerRange = text.textMarkerRangeForElement(text);
    startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
    currentMarker = advanceAndVerify(startMarker, 1, text);
    currentMarker = advanceAndVerify(currentMarker, 15, text);
    
    // Check getting the correct sentences.
    text = accessibilityController.accessibleElementById("text5");
    textMarkerRange = text.textMarkerRangeForElement(text);
    startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);

    // FIXME: Add in verifySentences() to match the original test once next/previous sentence
    // are implemented.

    debug(output);
    
    function advanceAndVerify(currentMarker, offset, obj) {
        var previousMarker = currentMarker;
        for (var i = 0; i < offset; i++) {
            previousMarker = currentMarker;
            currentMarker = obj.nextTextMarker(previousMarker);
        }
        verifySentenceRangeForTextMarker(previousMarker, currentMarker, obj);
        output += "\n";
        return currentMarker;
    }
    
    function replaceAttachmentInString(str) {
        str =  str.replace(String.fromCharCode(65532), "[ATTACHMENT]");
        return str;
    }
    
    function verifySentenceRangeForTextMarker(preMarker, textMarker, obj) {
        var markerRange = obj.textMarkerRangeForMarkers(preMarker, textMarker);
        var currentCharacter = replaceAttachmentInString(obj.stringForTextMarkerRange(markerRange));
        output += `Current character is: ${currentCharacter}\n`;
        
        var range = obj.sentenceTextMarkerRangeForTextMarker(textMarker);
        var sentence = replaceAttachmentInString(obj.stringForTextMarkerRange(range));
        output += `Current sentence is: ${sentence}\n`;
        
        // FIXME: Add "Pre sentence start to next sentence end" to match original test once 
        // previousSentenceStartTextMarkerForTextMarker and nextSentenceEndTextMarkerForTextMarker
        // are implemented off the main thread.
    }
}
</script>

</body>
</html>
